rename/#52 퀘스트 에러 타입 quest usecase 폴더 내부로 이동
