+++ application/usecases/quest/errors/CompleteQuestError.ts
